More VS tweaks
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1160 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/wc3270/wizard.c
@@ -322,7 +322,7 @@ static km_t *
 save_keymap_name(char *path, char *keymap_name, char *description)
 {
 	km_t *km;
-    	int sl;
+    	size_t sl;
 	km_t *kms;
 	FILE *f;
 	enum { KMF_BOTH, KMF_3270, KMF_NVT } km_mode = KMF_BOTH;
